Extract pending-prefix continuation logic into its own module
`src/wrap.rs` had grown to 447 lines, breaching the 400-line ceiling
documented in `AGENTS.md`. Most of the excess was the
join/update/dispatch state machine that reconciles a soft-wrapped
continuation chunk with the active `PendingPrefix` buffer.

Move that state machine into a new `wrap::continuation` submodule
without behaviour change: `apply_continuation_chunk`,
`update_span_state`, `split_reopen_span`, the `SpanStateUpdate` enum,
and the supporting helpers (`join_pending_continuation`,
`continuation_needs_leading_space`, `emit_pending_prefix_segment`)
all relocate there. `handle_pending_continuation` in `wrap.rs` keeps
its general dispatch role and calls `apply_continuation_chunk`
through a `pub(super)` re-export.

The new module borrows `PendingPrefix`, `PrefixLine`,
`ParagraphState`, and `ParagraphWriter` via the existing `pub(super)`
visibility on `wrap::paragraph`; tokenize helpers (`parse_open_code_span`,
`position_after_close`, `scan_continuation_span_state`,
`opening_fence_run_len`) are accessed via `super::tokenize`. No public
or `pub(crate)` surface changes.

`wrap.rs` is now 246 lines and `continuation.rs` is 219 lines.
